requirements/common.txt
django==3.0
djangorestframework~=3.11.0
django-extensions~=2.2.6
django-filter==2.2.0
drf-yasg==1.17.1
